+++ example/lib/main.dart
@@ -2,10 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 
-import 'package:ramp_flutter/configuration.dart';
-import 'package:ramp_flutter/host_event.dart';
 import 'package:ramp_flutter/ramp_flutter.dart';
-import 'package:ramp_flutter/widget_event.dart';
 
 import 'secrets.dart';