+++ lib/ramp_flutter.dart
@@ -4,6 +4,10 @@ import 'package:ramp_flutter/host_event.dart';
 import 'package:ramp_flutter/src/ramp_webview.dart';
 import 'package:ramp_flutter/widget_event.dart';
 
+export 'package:ramp_flutter/configuration.dart';
+export 'package:ramp_flutter/host_event.dart';
+export 'package:ramp_flutter/widget_event.dart';
+
 class RampFlutter {
   RampFlutter(Uri widgetUrl) : _webView = RampWebView(widgetUrl);